Fix: Responsive header
